chore: refactor csv on service
