+++ packages/backend/src/routers/projectRouter.ts
@@ -35,6 +35,7 @@ import {
 import Logger from '../logger';
 import { CsvService } from '../services/CsvService/CsvService';
 import {
+    csvService,
     dashboardService,
     projectService,
     s3Service,
@@ -343,37 +344,18 @@ projectRouter.post(
     allowApiKeyAuthentication,
     isAuthenticated,
     async (req, res, next) => {
-        const jobId = nanoid();
-
-        const baseAnalyticsProperties: DownloadCsv['properties'] = {
-            jobId,
-            userId: req.user?.userUuid,
-            organizationId: req.user?.organizationUuid,
-            projectId: req.params.projectUuid,
-            fileType: 'csv',
-            storage: s3Service.isEnabled() ? 's3' : 'local',
-        };
+        const { body } = req;
 
         try {
-            const { organizationUuid } = req.user!;
-            const { projectUuid } = req.params;
-            if (
-                req.user!.ability.cannot(
-                    'manage',
-                    subject('ExportCsv', { organizationUuid, projectUuid }),
-                )
-            ) {
-                throw new ForbiddenError();
-            }
-
-            const { body } = req;
             const {
-                csvLimit,
                 onlyRaw,
+                csvLimit,
                 showTableNames,
                 customLabels,
                 columnOrder,
             } = body;
+            const { projectUuid, exploreId } = req.params;
+
             const metricQuery: MetricQuery = {
                 dimensions: body.dimensions,
                 metrics: body.metrics,
@@ -384,74 +366,17 @@ projectRouter.post(
                 additionalMetrics: body.additionalMetrics,
             };
 
-            const numberColumns =
-                metricQuery.dimensions.length +
-                metricQuery.metrics.length +
-                metricQuery.tableCalculations.length;
-            const analyticsProperties: DownloadCsv['properties'] = {
-                ...baseAnalyticsProperties,
-                tableId: req.params.exploreId,
-                values: onlyRaw ? 'raw' : 'formatted',
-                context: 'results',
-                limit: parseAnalyticsLimit(csvLimit),
-
-                numColumns: numberColumns,
-            };
-            analytics.track({
-                event: 'download_results.started',
-                userId: req.user?.userUuid!,
-                properties: analyticsProperties,
-            });
-
-            const rows = await projectService.runQuery(
-                req.user!,
+            const fileUrl = await csvService.downloadCsv({
+                user: req.user!,
+                projectUuid,
+                exploreId,
                 metricQuery,
-                req.params.projectUuid,
-                req.params.exploreId,
-                csvLimit,
-            );
-            const numberRows = rows.length;
-
-            const explore = await projectService.getExplore(
-                req.user!,
-                req.params.projectUuid,
-                req.params.exploreId,
-            );
-            const itemMap = getItemMap(
-                explore,
-                metricQuery.additionalMetrics,
-                metricQuery.tableCalculations,
-            );
-
-            const fileId = await CsvService.convertRowsToCsv(
-                rows,
                 onlyRaw,
-                metricQuery,
-                itemMap,
+                csvLimit,
                 showTableNames,
                 customLabels,
-                columnOrder || [],
-            );
-
-            let fileUrl;
-            try {
-                const csvContent = fs.createReadStream(`/tmp/${fileId}`);
-                fileUrl = await s3Service.uploadCsv(csvContent, fileId);
-
-                await fsPromise.unlink(`/tmp/${fileId}`);
-            } catch (e) {
-                fileUrl = `${lightdashConfig.siteUrl}/api/v1/projects/${req.params.projectUuid}/csv/${fileId}`;
-            }
-
-            analytics.track({
-                event: 'download_results.completed',
-                userId: req.user?.userUuid!,
-                properties: {
-                    ...analyticsProperties,
-                    numRows: numberRows,
-                },
+                columnOrder,
             });
-
             res.json({
                 status: 'ok',
                 results: {
@@ -459,15 +384,6 @@ projectRouter.post(
                 },
             });
         } catch (e) {
-            analytics.track({
-                event: 'download_results.error',
-                userId: req.user?.userUuid!,
-                properties: {
-                    ...baseAnalyticsProperties,
-                    error: `${e}`,
-                },
-            });
-
             next(e);
         }
     },
@@ -846,77 +762,16 @@ projectRouter.post(
     allowApiKeyAuthentication,
     isAuthenticated,
     async (req, res, next) => {
-        const jobId = nanoid();
-
-        const analyticsProperties: DownloadCsv['properties'] = {
-            jobId,
-            userId: req.user?.userUuid,
-            organizationId: req.user?.organizationUuid,
-            projectId: req.params.projectUuid,
-            tableId: 'sql_runner',
-            values: 'raw',
-            context: 'sql runner',
-            fileType: 'csv',
-            storage: s3Service.isEnabled() ? 's3' : 'local',
-        };
         try {
-            const { organizationUuid } = req.user!;
+            const { customLabels, sql } = req.body;
             const { projectUuid } = req.params;
 
-            if (
-                req.user!.ability.cannot(
-                    'manage',
-                    subject('ExportCsv', { organizationUuid, projectUuid }),
-                )
-            ) {
-                throw new ForbiddenError();
-            }
-
-            analytics.track({
-                event: 'download_results.started',
-                userId: req.user?.userUuid!,
-                properties: {
-                    ...analyticsProperties,
-                },
+            const fileUrl = await csvService.downloadSqlCsv({
+                user: req.user!,
+                projectUuid,
+                sql,
+                customLabels,
             });
-
-            const results: ApiSqlQueryResults =
-                await projectService.runSqlQuery(
-                    req.user!,
-                    req.params.projectUuid,
-                    req.body.sql,
-                );
-
-            const csvContent = await CsvService.convertSqlQueryResultsToCsv(
-                results,
-                req.body.customLabels,
-            );
-
-            const fileId = `csv-${jobId}.csv`;
-
-            let fileUrl;
-            try {
-                fileUrl = await s3Service.uploadCsv(csvContent, fileId);
-            } catch (e) {
-                // Can't store file in S3, storing locally
-                await fsPromise.writeFile(
-                    `/tmp/${fileId}`,
-                    csvContent,
-                    'utf-8',
-                );
-                fileUrl = `${lightdashConfig.siteUrl}/api/v1/projects/${req.params.projectUuid}/csv/${fileId}`;
-            }
-
-            analytics.track({
-                event: 'download_results.completed',
-                userId: req.user?.userUuid!,
-                properties: {
-                    ...analyticsProperties,
-                    numRows: results.rows.length,
-                    numColumns: Object.keys(results.fields).length,
-                },
-            });
-
             res.json({
                 status: 'ok',
                 results: {
@@ -924,15 +779,6 @@ projectRouter.post(
                 },
             });
         } catch (e) {
-            analytics.track({
-                event: 'download_results.error',
-                userId: req.user?.userUuid!,
-                properties: {
-                    ...analyticsProperties,
-                    error: `${e}`,
-                },
-            });
-
             next(e);
         }
     },
